remove legacy manager
