+++ Engine/Source/renderer/src/assets.rs
@@ -1,54 +1,112 @@
-mod manager;
-
-use std::collections::HashMap;
-
 use ash::vk;
+use std::collections::HashMap;
+use std::marker::PhantomData;
 
 use crate::{
     Result,
     resources::{
         buffer::{IndexBuffer, VertexBuffer},
-        device::RenderDevice,
+        device::{Garbage, RenderDevice},
         image::Image,
     },
     utils::Vertex,
 };
 
-/// Complete GPU model.
-pub struct Model {
-    pub name: String,
-    pub primitives: Vec<Primitive>,
-    pub textures: Vec<Texture>,
-    pub materials: Vec<resource_manager::Material>,
-    /// One descriptor set per entry in `materials`.
-    /// `vk::DescriptorSet::null()` if the material has no base-colour texture.
-    pub material_sets: Vec<vk::DescriptorSet>,
+#[derive(Debug, PartialEq, Eq, Hash)]
+pub struct Handle<T> {
+    id: u32,
+    _marker: PhantomData<T>,
 }
 
-pub use manager::Texture;
+impl<T> Copy for Handle<T> {}
+impl<T> Clone for Handle<T> {
+    fn clone(&self) -> Self {
+        Self {
+            id: self.id,
+            _marker: PhantomData,
+        }
+    }
+}
 
-/// GPU-side handles for a single glTF primitive.
+pub struct Texture {
+    pub device: RenderDevice,
+    pub image: Image,
+    pub sampler: vk::Sampler,
+    pub mip_levels: u32,
+}
+
+impl Drop for Texture {
+    fn drop(&mut self) {
+        self.device.destroy(Garbage::Sampler(self.sampler));
+    }
+}
+
 pub struct Primitive {
-    pub vertex_buffer: VertexBuffer,
-    pub index_buffer: IndexBuffer,
-    pub index_count: u32,
-    pub material: Option<usize>,
+    vertex_buffer: VertexBuffer,
+    index_buffer: IndexBuffer,
+    index_count: u32,
+    material_handle: Option<Handle<Material>>,
+}
+
+pub struct Mesh {
+    primitives: Vec<Primitive>,
+}
+
+pub struct Material {
+    base_color: Handle<Texture>,
+    descriptor_set: vk::DescriptorSet,
+}
+
+pub struct Model {
+    // TODO: store transform with each mesh handle
+    pub mesh_instances: Vec<Handle<Mesh>>,
+}
+
+/// TODO: better storage type for assets
+struct AssetStorage<T> {
+    assets: Vec<Option<T>>,
+}
+
+impl<T> AssetStorage<T> {
+    fn new() -> Self {
+        Self { assets: Vec::new() }
+    }
+    fn insert(&mut self, asset: T) -> Handle<T> {
+        let id = self.assets.len() as u32;
+        self.assets.push(Some(asset));
+        Handle {
+            id,
+            _marker: PhantomData,
+        }
+    }
+    fn get(&self, handle: Handle<T>) -> &T {
+        self.assets[handle.id as usize]
+            .as_ref()
+            .expect("Invalid Handle")
+    }
+}
+
+pub struct AssetManager {
+    device: RenderDevice,
+
+    textures: AssetStorage<Texture>,
+    meshes: AssetStorage<Mesh>,
+    materials: AssetStorage<Material>,
+    models: AssetStorage<Model>,
+
+    path_to_model: HashMap<String, Handle<Model>>,
+
+    material_pool: vk::DescriptorPool,
 }
 
 /// TODO: also find a way to do this dynamically
 const MAX_MATERIAL_DESCRIPTOR_SET: u32 = 256;
 
-pub struct AssetManager {
-    device: RenderDevice,
-    models: HashMap<String, Model>,
-
-    material_descriptor_pool: vk::DescriptorPool,
-}
+const MODELS_PATH: &str = env!("MODELS_PATH");
 
 impl AssetManager {
-    pub fn new(device: &RenderDevice) -> Result<Self> {
-        // MATERIAL DESCRIPTOR SETS
-        let material_descriptor_pool = {
+    pub fn new(device: RenderDevice) -> Result<Self> {
+        let material_pool = {
             let pool_size = vk::DescriptorPoolSize {
                 ty: vk::DescriptorType::COMBINED_IMAGE_SAMPLER,
                 descriptor_count: MAX_MATERIAL_DESCRIPTOR_SET,
@@ -59,41 +117,65 @@ impl AssetManager {
 
             unsafe { device.device.create_descriptor_pool(&pool_info, None)? }
         };
+
         Ok(Self {
-            device: device.clone(),
-            models: HashMap::new(),
-            material_descriptor_pool,
+            device,
+            textures: AssetStorage::new(),
+            meshes: AssetStorage::new(),
+            materials: AssetStorage::new(),
+            models: AssetStorage::new(),
+            path_to_model: HashMap::new(),
+            material_pool,
         })
     }
-    pub fn get_or_upload_model(&mut self, name: &str) -> Result<&Model> {
-        if self.models.contains_key(name) {
-            return Ok(self.models.get(name).unwrap());
+
+    pub fn load_model(&mut self, name: &str) -> Result<Handle<Model>> {
+        if let Some(handle) = self.path_to_model.get(name) {
+            return Ok(*handle);
         }
 
-        self.upload_model(resource_manager::ResourceManager::load_model(name)?)
-    }
-    pub fn upload_model(&mut self, model: resource_manager::Model) -> Result<&Model> {
-        let primitives = model
+        let (gltf, buffers, images) = gltf::import(Self::model_path(name))?;
+
+        let texture_handles = images
+            .iter()
+            .map(|image| {
+                let tex = Image::upload_texture(self.device.clone(), image)?;
+                Ok(self.textures.insert(tex))
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        let material_handles =
+            self.create_materials(gltf.materials().collect(), &texture_handles)?;
+
+        let mesh_instances = gltf
             .meshes()
-            .iter()
-            .flat_map(|m| m.primitives().iter())
-            .map(|p| self.upload_primitive(p))
-            .collect::<Result<_>>()?;
+            .map(|mesh| {
+                let primitives = mesh
+                    .primitives()
+                    .map(|prim| Ok(self.upload_primitive(prim, &buffers, &material_handles)?))
+                    .collect::<Result<Vec<_>>>()?;
+                Ok(self.meshes.insert(Mesh { primitives }))
+            })
+            .collect::<Result<Vec<_>>>()?;
 
-        let textures: Vec<_> = model
-            .textures()
-            .iter()
-            .map(|t| Image::upload_texture(self.device.clone(), t))
-            .collect::<Result<_>>()?;
+        let model_handle = self.models.insert(Model { mesh_instances });
+        self.path_to_model.insert(name.to_string(), model_handle);
+        Ok(model_handle)
+    }
 
-        let material_count = model.materials().len();
+    fn create_materials(
+        &mut self,
+        materials: Vec<gltf::Material>,
+        texture_refs: &[Handle<Texture>],
+    ) -> Result<Vec<Handle<Material>>> {
+        let material_count = materials.len();
         // Allocate one set per material
         let layouts: Vec<vk::DescriptorSetLayout> =
             vec![self.device.layouts.material; material_count];
 
         let material_sets: Vec<vk::DescriptorSet> = if material_count > 0 {
             let alloc_info = vk::DescriptorSetAllocateInfo::default()
-                .descriptor_pool(self.material_descriptor_pool)
+                .descriptor_pool(self.material_pool)
                 .set_layouts(&layouts);
 
             unsafe { self.device.device.allocate_descriptor_sets(&alloc_info)? }
@@ -101,61 +183,93 @@ impl AssetManager {
             Vec::new()
         };
 
-        // Write the base-colour sampler into each set that has one
-        for (i, mat) in model.materials().iter().enumerate() {
-            let Some(&tex_idx) = mat.base_color_texture().as_ref() else {
-                continue; // leave this set in its default (null) state
-            };
+        Ok(materials
+            .into_iter()
+            .enumerate()
+            .map(|(i, mat)| {
+                let pbr = mat.pbr_metallic_roughness();
+                // TODO: actually PBR materials
+                let tex = pbr.base_color_texture().unwrap().texture().source().index();
 
-            let tex = &textures[tex_idx];
+                let tex_handle = texture_refs[tex];
+                let tex = self.textures.get(tex_handle);
+                let image_info = vk::DescriptorImageInfo::default()
+                    .image_layout(vk::ImageLayout::SHADER_READ_ONLY_OPTIMAL)
+                    .image_view(tex.image.view())
+                    .sampler(tex.sampler);
 
-            let image_info = vk::DescriptorImageInfo::default()
-                .image_layout(vk::ImageLayout::SHADER_READ_ONLY_OPTIMAL)
-                .image_view(tex.image.view())
-                .sampler(tex.sampler);
+                let write = vk::WriteDescriptorSet::default()
+                    .dst_set(material_sets[i])
+                    .dst_binding(2) // matches layouts.material
+                    .descriptor_type(vk::DescriptorType::COMBINED_IMAGE_SAMPLER)
+                    .image_info(std::slice::from_ref(&image_info));
 
-            let write = vk::WriteDescriptorSet::default()
-                .dst_set(material_sets[i])
-                .dst_binding(2) // matches layouts.material
-                .descriptor_type(vk::DescriptorType::COMBINED_IMAGE_SAMPLER)
-                .image_info(std::slice::from_ref(&image_info));
-
-            unsafe { self.device.device.update_descriptor_sets(&[write], &[]) };
-        }
-
-        self.models.insert(
-            model.name().to_string(),
-            Model {
-                name: model.name().to_owned(),
-                primitives,
-                textures,
-                materials: model.materials().to_vec(),
-                material_sets,
-            },
-        );
-        Ok(self.models.get(model.name()).unwrap())
+                unsafe { self.device.device.update_descriptor_sets(&[write], &[]) };
+                self.materials.insert(Material {
+                    base_color: tex_handle,
+                    descriptor_set: material_sets[i],
+                })
+            })
+            .collect())
     }
 
-    fn upload_primitive(&mut self, prim: &resource_manager::Primitive) -> Result<Primitive> {
-        let vertices: Vec<Vertex> = prim
-            .positions()
+    fn upload_primitive(
+        &self,
+        primitive: gltf::Primitive,
+        buffers: &[gltf::buffer::Data],
+        mat_refs: &[Handle<Material>],
+    ) -> Result<Primitive> {
+        let reader = primitive.reader(|buf| Some(&buffers[buf.index()]));
+
+        let positions: Vec<_> = reader
+            .read_positions()
+            .map(Iterator::collect)
+            .unwrap_or_default();
+        let normals: Vec<_> = reader
+            .read_normals()
+            .map(Iterator::collect)
+            .unwrap_or_default();
+        let texcoords: Vec<_> = reader
+            .read_tex_coords(0)
+            .map(|iter| iter.into_f32().collect())
+            .unwrap_or_default();
+        let indices: Vec<_> = reader
+            .read_indices()
+            .map(|iter| iter.into_u32().collect())
+            .unwrap_or_default();
+
+        let vertices: Vec<Vertex> = positions
             .iter()
             .enumerate()
             .map(|(i, &position)| Vertex {
                 position,
-                normal: prim.normals().get(i).copied().unwrap_or([0.0, 1.0, 0.0]),
-                texcoord: prim.texcoords().get(i).copied().unwrap_or([0.0, 0.0]),
+                normal: normals.get(i).copied().unwrap_or([0.0, 1.0, 0.0]),
+                texcoord: texcoords.get(i).copied().unwrap_or([0.0, 0.0]),
             })
             .collect();
 
         let vertex_buffer = VertexBuffer::upload_slice(self.device.clone(), &vertices)?;
-        let index_buffer = IndexBuffer::upload_slice(self.device.clone(), prim.indices())?;
+        let index_buffer = IndexBuffer::upload_slice(self.device.clone(), &indices)?;
 
         Ok(Primitive {
             vertex_buffer,
             index_buffer,
-            index_count: prim.indices().len() as u32,
-            material: *prim.material(),
+            index_count: indices.len() as u32,
+            material_handle: primitive.material().index().map(|idx| mat_refs[idx]),
         })
+    }
+
+    pub fn get_mesh(&self, handle: Handle<Mesh>) -> &Mesh {
+        self.meshes.get(handle)
+    }
+
+    pub fn get_material(&self, handle: Handle<Material>) -> &Material {
+        self.materials.get(handle)
+    }
+
+    /// Gives the path to a model from its name.
+    /// Just a simple utility to get `MODELS_PATH/{name}/{name}.gltf`.
+    fn model_path(name: &str) -> String {
+        format!("{MODELS_PATH}/{name}/{name}.gltf")
     }
 }
